requirements.txt
customtkinter
psycopg2-binary
python-dotenv
bcrypt